Refactoring

+++ website/lib/scenarios/delete.ts
@@ -1,14 +1,17 @@
 "use server";
 
+import prisma from "@/lib/prisma";
 import { revalidatePath } from "next/cache";
-import prisma from "../prisma";
 
 export async function deleteUserScenario(formData: FormData) {
-  const userScenarioId = formData.get("userScenarioId") as string;
+  const userScenarioId = formData.get("userScenarioId");
+  if (typeof userScenarioId !== "string" || !userScenarioId)
+    throw new Error("Missing userScenarioId");
+
   await prisma.userScenario.delete({
     where: {
       id: userScenarioId,
     },
   });
-  revalidatePath("/admin/users"); 
+  revalidatePath("/admin/users");
 }
